Bug fixes
 - Prevent valgrind from complaining about tinymt
 - stop patching builtins when done bootstrapping
 - NativePatch pointers were being invalidated
 - add source_id() func
 - add List.flatMap, change List.flatten, change Stack.set_env & co
 - Make sure load_module() is idempotent
 - catch division by 0. Still need to improve error reporting
 - add Vec2.distance_sqr and magnitude_sqr
 - Bring back performance stat tracking

+++ src/file.cpp
@@ -208,7 +208,7 @@ void read_text_file(const char* filename, caValue* contentsOut)
     size_t bytesRead = fread(as_blob(contentsOut), 1, file_size, fp);
     
     if (bytesRead < file_size)
-        string_resize(contentsOut, bytesRead);
+        string_resize(contentsOut, (int) bytesRead);
 
     fclose(fp);
 }
